feat: add CLI application with basic configuration handling, and access token authentication (#1)
## Description of the change

Added a basic CLI application based on cobra and viper libraries. On top
of the default cobra functionalities (help, version), it also supports
the following:
- rudder-cli auth login - asks for an access token and stores in the
default configuration file
- rudder-cli debug config
- rudder-cli debug config-file - used for debugging purposes. Dumps
active configuration and config file path.

Debugging needs a `"debug": true` field in the configuration JSON.
Default configuration location is `~/.rudder/config.json` and is created
by default if it's missing.

## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request