feat: add support for configuring rudderstack api url (#8)
## Description of the change

This introduces the following changes:
- A new optional field `apiURL` can be added to the CLI's config file,
in case it needs to access a different RudderStack API URL (e.g
staging).
- The `GetAccessToken()` function in the `config` package has been
replaced by a `GetConfig()` function which returns the entire effective
config as a struct. Going forward, that should be the preferred way of
exposing configurations to various packages, instead of implementing
dedicated getters.
- Configuring both access token and API url is now possible through the
`RUDDERSTACK_ACCESS_TOKEN` and `RUDDERSTACK_API_URL` environmental
variables. The names match the same variables that would be used to
configure the same for our terraform provider.

## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request