feat: added support to check if error is not found and allow the operation to succeed if that's the case (#25)
## Description of the change
In order to not get stuck permanently with the cli, we would need to
handle the situation where the cli is trying to remove an entity already
removed from upstream. This generates an error on the catalog client but
we would need to silence the error and move forward.

If not done, customer state could be in limbo and would require
unnecessary manual intervention.


## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request