fix: docker images are built on tag pushes, instead of releases (#56)
## Description of the change

This is because GitHub doesn't trigger the docker workflow from the
actions of another (goreleaser) flow.

From GitHub documentation:
> When you use the repository's GITHUB_TOKEN to perform tasks, events
triggered by the GITHUB_TOKEN, with the exception of workflow_dispatch
and repository_dispatch, will not create a new workflow run. This
prevents you from accidentally creating recursive workflow runs. For
example, if a workflow run pushes code using the repository's
GITHUB_TOKEN, a new workflow will not run even when the repository
contains a workflow configured to run when push events occur. For more
information, see Automatic token authentication.

## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request