feat(cli): retl provider for sql models (#114)
## Description of the change
The pr adds a retl provider. The provider depends on handlers. As a
first approach only the SQL model handler is added. This handler exposes
the functionality to update/create/remove a sql model from the platform.

The pr also adds a generic command for apply, validate, and destroy
```
rudder-cli apply -l test_projects
rudder-cli validate -l test_projects
rudder-cli destroy
```
Implemented tickets
-
[WAR-793](https://linear.app/rudderstack/issue/WAR-793/rudder-cli-implement-retl-provider)
-
[WAR-790](https://linear.app/rudderstack/issue/WAR-790/rudder-cli-implement-generic-project-commands)


## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request