fix: remove unnecessary UnmarshalJSON when handling variantcase and updated the match value validation (#144)
## Description of the change

Unmarshalling the variantcase in a custom way, made sure that the
integer value can be properly handled in the match. The consequence of
that handling was, we started to get unintentional diffs even when there
was no change in the catalog.

The main reason was the type of match value we applied was `int` due to
custom unmarshal JSON while the value read from the state is still
`float64` which is default type of reading numbers by Go. This generates
unnecessary diffs everytime without any changes.

## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request