feat: implement custom differ for custom types to prevent unnecessary updates (#150)
## Description of the change

When adding variants onto the custom types model, we identified that the
state differ started showing empty diffs on the custom types and
trackingplans. Leave unchecked, they made API call to update the custom
types with the same datapoint as it is. This unecessarily bumped the
trackingplan versioning and hence creates hassle.

This PR implements a custom differ on the custom-types to prevent any
unnecessary API updates.

## Type of change
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Related issues

> Fix [#1]() 

## Checklists

### Development

- [ ] Lint rules pass locally
- [ ] The code changed/added as part of this pull request has been
covered with tests
- [ ] All tests related to the changed code pass in development

### Code review 

- [ ] This pull request has a descriptive title and information useful
to a reviewer. There may be a screenshot or screencast attached
- [ ] "Ready for review" label attached to the PR and reviewers
mentioned in a comment
- [ ] Changes have been reviewed by at least one other engineer
- [ ] Issue from task tracker has a link to this pull request