feat: separate out the importable resources in the print diff (#201)
## Description
We are clubbing importable resources with new resources when diffing the
previous and current state. This creates confusion for the user as to
what resources are getting imported and what will be created , updated /
removed.

## Changes
1. Added a new `importableResources` field on the differ
2. Using the import metadata on the resource with current workspaceId as
indication to add it to the current importable resources list.
3. Created a new `DiffOptions` to cleanly pass WorkspaceID to the
differ.

## Example
<img width="674" height="194" alt="Screenshot 2025-10-04 at 12 12 34 AM"
src="https://github.com/user-attachments/assets/f4b06eae-7110-45ff-bb62-32f6ca2f64a1"
/>