feat: prevent importing workspace when detected changes not synced (#219)
## Description
Importing workspace require us to know before hand all the entities in
the project as we use it to feed the namer module. Given the project is
not synced to upstream could mean that we are missing deleted entries
being added to the resource graph.

This could mean we can now have the namer module provide `externalIDs`
which clashes with the deleted entries.

To prevent a case like this and other edge cases, we propose to
`ComputeDiff` first and only when we don't have any diff, move ahead
with importing workspace entities.

## Changes
To keep the changes minimal:
1. `LoadState` from the provider which depicts the current state of the
resources.
2. Once we have a state, we convert it into a source graph.
3. Detect changes from the sourceGraph to target graph already generated
within the provider.
4. Only import when no diff detected.