fix: make property type as optional in validation (#223)
## Description
We are making the type as not mandatory in the CLI as upstream data
catalog allows user creating the property with empty types. This problem
gets exacerbated when we have import command, allowing and rendering
those imports but the validate command run immediately after during the
apply fails.

This PR fixes this by relaxing the rules on type being allowed to be
empty and validate command passing in case found a property without a
type.