fix: set identitySection to properties by default if it is not explicitly set in the project YAMLs (#217)
Re-applying a project should not show any diffs.
When "identifySection" is omitted in the tracking plan YAML, re-applying
shows a diff for this field. The diff indicates the current/remote
value is "properties" while the expected value is an empty string ('').
This PR fixes the above by setting the identitySection to properties by
default if it is not explicitly set in the project YAMLs