feat: remove the constraints on the ordering of the event states (#292)
## Description
We have now moved away from the localID as the connecting identifier for
calculating diffs in the trackingplan to the upstream identifier of the
events and properties. This results in a less fragile ecosystem for diff
comparison in the trackingplans.

## Changes
1. Used event upstream identifier as the key for lookup events on the
trackingplans
2. Used property upstream identifier as the key of lookup properties on
the trackingplan events.
3. Stopped capturing the event states on the responses to populate the
event identifier. `This improvement then becomes the bed on which we can
remove the need to rebuild the schema on every upsert call as we don't
depend on the response ordering`. After this we can get this change
https://github.com/rudderlabs/rudder-iac/pull/286 merged