feat: batched trackingplan event update requests (#308)
## Description
A bottleneck which we identified when applying trackingplan with large
number of event rules is the single payload acceptance in the
trackingplan update request. Given we know have backend allowing batched
updates to the event rules, we are opening up the updates to the event
rules to be batched now.

In addition to this, to handle existing tech debt, we have now:
1. Improved the function naming in Trackingplan Store to bring inline
implementation of function with it's name.
2. Consolidated the data model in the API
3. Make functions private if they are not used outside.