feat: use rebuildSchemas=false to fetch latest mappings (#286)
## Description
In order to fetch the event identifiers on the trackingplan without
rebuilding the schemas, we need to pass an additional query param called
`rebuildSchemas=false`. This query param helps in pulling the latest
mappings from the db without the need to explicitly start rebuilding the
schemas.

In previous cli versions, we needed this capability but going forward we
don't require it so to be backwards compatible, we need to explicitly
pass the query param so older clients can work properly.

## Changes
1. Added ability of the `GetTrackingPlan` and `GetTrackingPlans`
function to specify if the rebuilding needs to happen now.
2. The actual API call to add the query param for this to the backend.
3. Remove the version on the trackingplan state given we are not
utilising the version.
4. The API client gets to keep the version as the typer uses it and
requires to make sure we pull the latest built version of the
trackingplan.