test: enable concurrent syncs e2e test (#313)
## Description
We are making the e2e tests run with concurrent syncs enable to save
test running e2e tests. This allows the test feedback loop to be faster.

In addition to this, we disabled the telemetry so we don't even send
time firing off a track event as it doesn't help during the e2e test.


## Results
e2e tests run **40%** faster


## Ticket

https://linear.app/rudderstack/issue/DAW-2668/enable-concurrent-syncs-in-e2e-tests