chore: allow setting additonalProperties for nested properties (#297)
This PR does the following:

1. Allows users to set the additionalProperties field for nested
properties.
If this field is set for a non-nested property, a validation error is
thrown.

2. Sets sane defaults for additionalProperties:
a. false for non-object types and non-nested properties
b. true for object-type properties
c. true for nested properties — this is already covered by (b), but we
are keeping it as it has always been for easier understanding

We avoided setting additionalProperties to true for all property types,
as this would have shown diffs for all user-defined properties during
the first apply (after this is released). We only set it to true where
the additionalProperties field is actually supported.